+++ misc/dbt-materialize/CHANGELOG.md
@@ -45,6 +45,8 @@
   ```
 * Upgrade to `dbt-postgres` v1.2.0.
 
+* Fully deprecate the custom index materialization.
+
 ## 1.1.3 - 2022-08-17
 
 * Deprecate the `mz_generate_name` macro. The native Jinja function [`{